make it DRYer
